@@ -8,7 +8,7 @@
     }
   ],
   "require": {
-    "php": ">=8.1",
+    "php": ">=8.3",
     "ext-json": "*",
     "ext-pdo": "*",
     "gcworld/common":"^2.7.0",
