@@ -9,33 +9,26 @@ use \Exception;
  */
 class ORMException extends Exception
 {
-    public $backtrace = null;
+    public readonly ?array $backtrace;
 
     /**
      * ORMException constructor.
-     * @param string          $message
-     * @param int             $code
-     * @param \Exception|null $previous
      */
     public function __construct(string $message, int $code = 0, Exception $previous = null)
     {
-        $this->backtrace = debug_backtrace();
         parent::__construct($message, $code, $previous);
+        $this->backtrace = debug_backtrace();
     }
 
     /**
      * custom string representation of object
-     * @return string
      */
-    public function __toString()
+    public function __toString(): string
     {
         return __CLASS__.": [{$this->code}]: {$this->message}\n";
     }
 
-    /**
-     * @return array|null
-     */
-    public function geTrace()
+    public function getTrace(): ?array
     {
         return $this->backtrace;
     }
